refactor(pipeline): modularize detection.py into submodules
- Extract Hybrid Consensus logic into detection_hybrid.py (HybridDetector class).
- Extract path resolution and VRAM utilities into detection_utils.py.
- Extract configuration constants into detection_config.py.
- Streamline DetectorOrchestrator in detection.py (reduced to ~120 lines).
- Update main.py and unit tests to reflect the new module structure.
- Functional logic remains unchanged, verified with unit and smoke tests.
